fix(web): hide authenticated app from search indexing
