+++ web/src/app/robots.ts
@@ -10,7 +10,24 @@ export default function robots(): MetadataRoute.Robots {
       {
         userAgent: "*",
         allow: "/",
-        disallow: ["/inboxes", "/api/", "/api-keys", "/domains", "/feedback"],
+        disallow: [
+          "/api/",
+          "/api-keys",
+          "/billing",
+          "/contacts",
+          "/dashboard",
+          "/domains",
+          "/feedback",
+          "/get-started",
+          "/inboxes",
+          "/metrics",
+          "/reviews",
+          "/settings",
+          "/templates",
+          "/trash",
+          "/webhook-secrets",
+          "/webhooks",
+        ],
       },
     ],
     sitemap: `${SITE_URL}/sitemap.xml`,
